requirements.txt
yt-dlp
mutagen
requests
pillow
